+++ src/1c4a/Error.ts
@@ -3,7 +3,6 @@ import { InternetmarkeError } from '../Error';
 export class OneClickForAppError extends InternetmarkeError {}
 
 export class PartnerError extends OneClickForAppError {}
-export class UserError extends OneClickForAppError {}
 
 export class AddressError extends OneClickForAppError {}
 export class CheckoutError extends OneClickForAppError {}